batch: Move line range view
realized_entries.py owns compact realized content and provenance-backed
entry views.

The _LineRange helper was a generic indexed slice view, and baseline
correspondence imported it from realized entries as private support. That
tied baseline matching to realized-entry internals.

This commit addresses that by moving the range view into line_range_view as
LineRangeView and routing realized entries plus baseline correspondence
through it.

Subsequent architecture coverage will record the public line-range view
boundary.
